requirements.txt
fastapi>=0.68.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiofiles>=0.7.0
websockets>=10.0
aiortc>=1.3.0
python-socketio>=5.4.0
websockets>=10.0
email-validator>=1.1.3
Pillow>=9.0.0